refactor: Update presentation and team repository methods for improved functionality - Modify presentation repository to include presentation_datetime and user_id in queries - Enhance team repository methods to handle soft deletes and transactions - Refactor team repository struct name for consistency - Update SQL initialization script to reflect changes in foreign key constraints and added fields

+++ backend/internal/repository/mysql/team.go
@@ -4,23 +4,23 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"github.com/SHIMA0111/PresentationManager/backend/internal/repository"
+
+	_ "github.com/go-sql-driver/mysql"
 
 	"github.com/SHIMA0111/PresentationManager/backend/internal/domain"
 )
 
-type TeamRepository struct {
+type teamRepository struct {
 	db *sql.DB
 }
 
-func NewTeamRepository(db *sql.DB) *TeamRepository {
-	return &TeamRepository{db: db}
+func NewTeamRepository(db *sql.DB) repository.TeamRepository {
+	return &teamRepository{db: db}
 }
 
-func (r *TeamRepository) CreateTeam(ctx context.Context, team *domain.Team) error {
-	query := `
-		INSERT INTO teams (id, name, description)
-		VALUES (?, ?, ?)
-	`
+func (r *teamRepository) CreateTeam(ctx context.Context, team *domain.Team) error {
+	query := "INSERT INTO teams (id, name, description) VALUES (?, ?, ?)"
 	_, err := r.db.ExecContext(ctx, query, team.Id, team.Name, team.Description)
 	if err != nil {
 		return fmt.Errorf("failed to insert team: %w", err)
@@ -29,11 +29,8 @@ func (r *TeamRepository) CreateTeam(ctx context.Context, team *domain.Team) erro
 	return nil
 }
 
-func (r *TeamRepository) AddTeamMember(ctx context.Context, teamMember *domain.TeamMember) error {
-	query := `
-		INSERT INTO team_members (team_id, user_id, Role)
-		VALUES (?, ?, ?)
-	`
+func (r *teamRepository) AddTeamMember(ctx context.Context, teamMember *domain.TeamMember) error {
+	query := "INSERT INTO team_members (team_id, user_id, Role) VALUES (?, ?, ?)"
 	_, err := r.db.ExecContext(ctx, query, teamMember.TeamId, teamMember.UserId, teamMember.Role)
 	if err != nil {
 		return fmt.Errorf("failed to insert team member: %w", err)
@@ -42,12 +39,9 @@ func (r *TeamRepository) AddTeamMember(ctx context.Context, teamMember *domain.T
 	return nil
 }
 
-func (r *TeamRepository) GetTeam(ctx context.Context, id string) (*domain.Team, error) {
-	query := `
-		SELECT id, name, description
-		FROM teams
-		WHERE id = ? AND deleted_at IS NULL
-	`
+func (r *teamRepository) GetTeam(ctx context.Context, id string) (*domain.Team, error) {
+	query := "SELECT id, name, description FROM teams WHERE id = ? AND deleted_at IS NULL"
+
 	var team domain.Team
 	if err := r.db.QueryRowContext(ctx, query, id).Scan(&team.Id, &team.Name, &team.Description); err != nil {
 		return nil, fmt.Errorf("failed to get team: %w", err)
@@ -56,12 +50,9 @@ func (r *TeamRepository) GetTeam(ctx context.Context, id string) (*domain.Team, 
 	return &team, nil
 }
 
-func (r *TeamRepository) GetTeams(ctx context.Context) ([]*domain.Team, error) {
-	query := `
-		SELECT id, name, description
-		FROM teams
-		WHERE deleted_at IS NULL
-	`
+func (r *teamRepository) GetTeams(ctx context.Context) ([]*domain.Team, error) {
+	query := "SELECT id, name, description FROM teams WHERE deleted_at IS NULL"
+
 	rows, err := r.db.QueryContext(ctx, query)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get teams: %w", err)
@@ -80,12 +71,9 @@ func (r *TeamRepository) GetTeams(ctx context.Context) ([]*domain.Team, error) {
 	return teams, nil
 }
 
-func (r *TeamRepository) UpdateTeam(ctx context.Context, team *domain.Team) error {
-	query := `
-		UPDATE teams
-		SET name = ?, description = ?
-		WHERE id = ? AND deleted_at IS NULL
-	`
+func (r *teamRepository) UpdateTeam(ctx context.Context, team *domain.Team) error {
+	query := "UPDATE teams SET name = ?, description = ? WHERE id = ? AND deleted_at IS NULL"
+
 	_, err := r.db.ExecContext(ctx, query, team.Name, team.Description, team.Id)
 	if err != nil {
 		return fmt.Errorf("failed to update team: %w", err)
@@ -94,38 +82,58 @@ func (r *TeamRepository) UpdateTeam(ctx context.Context, team *domain.Team) erro
 	return nil
 }
 
-func (r *TeamRepository) RemoveTeamMember(ctx context.Context, teamMember *domain.TeamMember) error {
-	query := `
-		DELETE FROM team_members
-		WHERE team_id = ? AND user_id = ?
-	`
-	_, err := r.db.ExecContext(ctx, query, teamMember.TeamId, teamMember.UserId)
+func (r *teamRepository) RemoveTeamMember(ctx context.Context, teamMember *domain.TeamMember) error {
+	tx, err := r.db.BeginTx(ctx, nil)
 	if err != nil {
+		return fmt.Errorf("failed to begin transaction: %w", err)
+	}
+	defer tx.Rollback()
+
+	presentationUpdateUserQuery := "UPDATE presentations SET user_id = NULL WHERE user_id = ? AND status != 3 AND deleted_at IS NULL"
+	if _, err = tx.ExecContext(ctx, presentationUpdateUserQuery, teamMember.UserId); err != nil {
+		return fmt.Errorf("failed to remove team member: %w", err)
+	}
+
+	teamMemberDeleteQuery := "DELETE FROM team_members WHERE team_id = ? AND user_id = ? AND deleted_at IS NULL"
+	if _, err = tx.ExecContext(ctx, teamMemberDeleteQuery, teamMember.TeamId, teamMember.UserId); err != nil {
 		return fmt.Errorf("failed to remove team member: %w", err)
 	}
 
 	return nil
 }
 
-func (r *TeamRepository) DeleteTeam(ctx context.Context, id string) error {
-	query := `
-		UPDATE teams
-		SET deleted_at = NOW()
-		WHERE id = ? AND deleted_at IS NULL
-	`
-	_, err := r.db.ExecContext(ctx, query, id)
+func (r *teamRepository) DeleteTeam(ctx context.Context, id string) error {
+	tx, err := r.db.BeginTx(ctx, nil)
 	if err != nil {
+		return fmt.Errorf("failed to begin transaction: %w", err)
+	}
+	defer tx.Rollback()
+
+	teamMemberDeleteQuery := "UPDATE team_members SET deleted_at = NOW() WHERE team_id = ? AND deleted_at IS NULL"
+	if _, err = tx.ExecContext(ctx, teamMemberDeleteQuery, id); err != nil {
 		return fmt.Errorf("failed to delete team: %w", err)
+	}
+
+	presentationUpdateUserQuery := "UPDATE presentations SET user_id = NULL, title = NULL, description = NULL WHERE user_id = ? AND status != 3 AND deleted_at IS NULL"
+	if _, err = tx.ExecContext(ctx, presentationUpdateUserQuery, id); err != nil {
+		return fmt.Errorf("failed to delete team: %w", err)
+	}
+
+	teamDeleteQuery := "UPDATE teams SET deleted_at = NOW() WHERE id = ? AND deleted_at IS NULL"
+	if _, err = tx.ExecContext(ctx, teamDeleteQuery, id); err != nil {
+		return fmt.Errorf("failed to delete team: %w", err)
+	}
+
+	if err = tx.Commit(); err != nil {
+		return fmt.Errorf("failed to commit transaction: %w", err)
 	}
 
 	return nil
 }
 
-func (r *TeamRepository) HardDeleteTeam(ctx context.Context, id string) error {
-	query := `
-		DELETE FROM teams
-		WHERE id = ?
-	`
+func (r *teamRepository) HardDeleteTeam(ctx context.Context, id string) error {
+	query := "DELETE FROM team_members WHERE team_id = ?"
+
 	_, err := r.db.ExecContext(ctx, query, id)
 	if err != nil {
 		return fmt.Errorf("failed to hard delete team: %w", err)
